adding synth preset editing, oscillator tuning, incremental updates to scale follower LUT generation

+++ src/clarinoid/loopstation/LooperHarmonizer.hpp
@@ -105,7 +105,7 @@ struct LooperAndHarmonizer
         mLayers[mCurrentlyWritingLayer].mIsPlaying = true;
 
         // and start with the new layer.
-        mCurrentlyWritingLayer = iNewLayer;
+        mCurrentlyWritingLayer = (uint8_t)iNewLayer;
         newLayer->StartRecording(mStatus, mv, buf, Ptr(EndPtr(LOOPSTATION_BUFFER)));
       }
